Add site_id to Participant and Site relationships
Added a new column 'site_id' to the 'participant' table. This change was required to form a many-to-one relationship between 'participant' and 'site' entities. Understanding which participant is affiliated with what site was needed in the application. Also, bi-directional methods in the Site and Participant entities were created to ensure easy access to the related objects.