Commit message: Reorganize table relationships and remove unnecessary Participant-Sortie link
Merged update and downgrade migration files into a single migration file and deleted the previous ones. These changes simplify the codebase and decrease the risk of errors by maintaining a single, consistent version of the database schema. Also, removed the many-to-many relationship between 'participant' and 'sortie' in the entity files as it is no longer needed in the application logic.