Add data fixtures and modify registration form
This commit introduces 'Doctrine Fixtures' to the project in order to standardise and simplify data loading. This includes definition of a 'Site' entity within the AppFixtures.php, which allows us to create and persist 'Site' instances for testing and development. In addition, the registration form is modified to include a 'Site' field that uses EntityType from symfony. The form is also styled with Bootstrap for a more cohesive user interface. The changes to the project will aid better data management and improve user experience during registration. The updated 'composer.lock' file indicates that the fixtures bundle is added to the list of project's dev dependencies.