Update login UI, enhance security and fix template bugs
This commit switches the login form from simple div layout to a more interactive form with fields to accept username and password. The change improves UI and user friendliness. An error handler is also added to display login errors.

Enhancements to the security is made by changing the security provider from the 'users_in_memory' to the 'app_user_provider' that fetches user details from the 'Participant' entity. This offers a more secure and realistic authentication process.

A minor fix in 'home/index.html.twig' template is done by renaming the variable 'sorties' to 'sortieList'. This was necessary because the name 'sorties' was causing confusion and was not compliant with best coding practice.

The registration template is also modified for a cleaner look and a logout link is added to 'base.html.twig' for better user navigation.

Changes made to 'LoginController.php' indicate that a route name has been corrected to link to the proper Controller method.