Updated form validation in ProfileType and fixed data setting in SortiesController
This commit includes multiple updates, mainly focusing on improving form validation and correcting certain data settings. In 'ProfileType' form, new constraints have been added to password fields and some fields are now required to improve form validation. In 'SortiesController', the way of checking if a 'sortie' has an ID is fixed and the process of setting certain 'sortie' properties is commented out to prevent data overriding.

Modification in 'SortieType' class introduces 'Site' selection with dropdown style improving the user interface. Change in 'Sortie.php' renames property 'Site' to 'site' for more accurate naming convention.

Some changes were made in 'ProfileController' to add a condition for updating password only if 'newPassword' is not null, thus preventing unnecessary database updates.

Additionally, the 'package-lock.json' is updated changing the project name to 'Projet'. This is likely a minor change, but it's important to keep track of all package-related updates.