Removed Etat entity and related dependency
This commit includes the deletion of the Etat and EtatRepository classes. This change was required due to a change in the data model where the 'Etat' feature was deemed unnecessary and thus removed.

The related migration file 'Version20231009122556.php' has also been updated to remove the respective lines for creating the 'etat' table and its relationship with the 'sortie' table.

This change simplifies the data model and reduces complexity in our codebase.