Change 'sortie' date field and introduce 'Etat' enumeration
Adjusted the 'Sortie' entity to use 'dateHeureDebut' instead of 'date' to display a more accurate timeframe. Also introduced 'Etat' enumeration, replacing the previous implementation.

A new library 'myclabs/php-enum' has been installed to help manage the enums. Adjustments were made to the 'SortieController' to use the newly created 'Etat' enum. Composer files and IDE settings have been updated accordingly for the new library.

The database schema was also updated to include 'etat' in the 'sortie' table via a migration.

Overall this change improves the accuracy of the 'sortie' timeframe and makes the management of 'etat' more straightforward and safer.