Update mappedBy annotation in Site entity
In the Site entity, the mappedBy attribute in the OneToMany annotation related to the Sortie class was updated from 'Site' to 'site'. This change was necessary to ensure the consistency of entity properties with annotations and to ensure correct database mappings for relationships between Site and Sortie entities. It will ensure accurate and expected results when joining or querying the related entities in the database.