Enable mail mapping in ProfileType and adjust UI layout
The 'mapped' attribute for the 'mail' field in the ProfileType form was uncommented to enable the mapping of this field. This change allows the 'mail' field to properly persist to the user profile.

Additionally, the layout of the 'editerSortie.html.twig' and 'index.html.twig' templates was adjusted. 'Photo de la Sortie' form group was removed for a streamlined UI design.

In the ProfileController, the UserIdentifier was added to provide the correct user context when rendering the profile page. This enables displaying the correct user's information in the profile form.