Update login page details, enhance events registration logic, modify project settings
Enhancements were made to the user's interaction with events on the home page, in particular regarding conditions for registering and cancelling registration for events. Change of ids on the login page for improved consistency. The 'Project' settings in .iml and package-lock.json files were updated and jQuery 3.6.0 library was added. Modifications were made to several Twig files and Symfony's security settings for improved usability. The final change was a substantial refactor in the SortiesController.php to improve the efficiency of code by removing unnecessary instructions and taking advantage of the if else structure in appropriate instances. These changes lead to better user experience and a more efficient and readable codebase.