Add cancellation form for sorties and refactor some pages
A new form, AnnulerSortieType has been introduced to handle the cancellation of a sortie. This comes with a corresponding template update. The status labels have been reassigned in Etat.php to provide a clearer understanding of the respective states of a sortie. Changes in home/index.html.twig, SortiesController.php, and others have also been made for better readability and usability. Modifications in VilleController.php now allow it to handle an optional 'id' parameter, providing a better user experience when editing towns.