"Refactor site form for improved readability
The form in the site's index file was restructured to improve readability and maintainability.

The nome form has been elevated to a separate div which now includes labels and better presentation through the usage of form classes. This change simplifies the form's appearance and makes it easier to maintain by leveraging Symfony's form functionalities."