"Add link to event view and display additional event details
To increase the application's usability, a link to the detailed view of each event was added directly from the event's name on the homepage.

Furthermore, the following additional information is now displayed for each event: the current state of the event, whether or not the currently logged in user is registered and the pseudonym of the event's organizer. This provides users will a fuller understanding of the event directly from the homepage."