"Update assertions for entity 'Sortie'"
Assertions for the entity 'Sortie' attributes have been updated. The assertions for 'nom' and 'infosSortie' were previously checking for a numeric range using 'GreaterThan' and 'LessThan'. However, these attributes are strings, therefore it would be more appropriate to use 'Length' assertion to enforce string length. This improves code coherence and accuracy of validations.